refactor: replace Gatsby with Vite
