--- packages/site/gatsby-browser.tsx
@@ -1,16 +1,0 @@
-import { GatsbyBrowser } from 'gatsby';
-import { StrictMode } from 'react';
-import { App } from './src/App';
-import { Root } from './src/Root';
-
-export const wrapRootElement: GatsbyBrowser['wrapRootElement'] = ({
-  element,
-}) => (
-  <StrictMode>
-    <Root>{element}</Root>
-  </StrictMode>
-);
-
-export const wrapPageElement: GatsbyBrowser['wrapPageElement'] = ({
-  element,
-}) => <App>{element}</App>;
