--- packages/site/gatsby-node.ts
@@ -1,11 +1,0 @@
-import type { GatsbyNode } from 'gatsby';
-
-export const onCreateWebpackConfig: GatsbyNode['onCreateWebpackConfig'] = ({
-  actions,
-}) => {
-  actions.setWebpackConfig({
-    experiments: {
-      asyncWebAssembly: true,
-    },
-  });
-};
