+++ packages/site/package.json
@@ -4,13 +4,13 @@
   "private": true,
   "license": "(MIT-0 OR Apache-2.0)",
   "scripts": {
-    "build": "cross-env GATSBY_TELEMETRY_DISABLED=1 gatsby build",
+    "build": "vite build",
     "clean": "rimraf public",
     "lint": "npm run lint:eslint && npm run lint:misc --check",
     "lint:eslint": "eslint . --cache --ext js,ts",
     "lint:fix": "npm run lint:eslint -- --fix && npm run lint:misc -- --write",
     "lint:misc": "prettier '**/*.json' '**/*.md' '!CHANGELOG.md' --ignore-path .gitignore",
-    "start": "cross-env GATSBY_TELEMETRY_DISABLED=1 gatsby develop",
+    "start": "vite",
     "test": "echo TODO"
   },
   "browserslist": {
@@ -34,7 +34,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-is": "^18.2.0",
-    "styled-components": "5.3.3"
+    "styled-components": "5.3.3",
+    "vite-plugin-wasm": "^3.2.2"
   },
   "devDependencies": {
     "@metamask/eslint-config": "^12.2.0",
@@ -52,6 +53,7 @@
     "@types/styled-components": "^5.1.25",
     "@typescript-eslint/eslint-plugin": "^5.33.0",
     "@typescript-eslint/parser": "^5.33.0",
+    "@vitejs/plugin-react": "^4.2.0",
     "cross-env": "^7.0.3",
     "dotenv": "^16.3.1",
     "eslint": "^8.55.0",
@@ -68,6 +70,8 @@
     "prettier": "^2.2.1",
     "prettier-plugin-packagejson": "^2.2.18",
     "rimraf": "^3.0.2",
-    "typescript": "^5.1.0"
+    "typescript": "^5.1.0",
+    "vite": "^4.4.12",
+    "vite-plugin-svgr": "^4.2.0"
   }
 }
