+++ packages/site/src/config/snap.ts
@@ -4,11 +4,11 @@
  */
 export const defaultSnapOrigin =
   // eslint-disable-next-line no-restricted-globals
-  process.env.SNAP_ORIGIN ?? `local:http://localhost:8080`;
+  import.meta.env.SNAP_ORIGIN ?? `local:http://localhost:8080`;
 
 export const getSnapParams = () => {
   // eslint-disable-next-line no-restricted-globals
-  const envSnapVersion = process.env.SNAP_VERSION;
+  const envSnapVersion = import.meta.env.SNAP_VERSION;
   if (defaultSnapOrigin.startsWith('npm:') && envSnapVersion) {
     return { version: envSnapVersion };
   }
