+++ packages/site/src/contexts/ToggleThemeContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export type ToggleTheme = () => void;
+
+export const ToggleThemeContext = createContext<ToggleTheme>(
+  (): void => undefined,
+);
