+++ packages/site/vite.config.ts
@@ -1,0 +1,8 @@
+import react from '@vitejs/plugin-react';
+import { defineConfig } from 'vite';
+import wasm from 'vite-plugin-wasm';
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react(), wasm()],
+});
